+++ IB-Tracker/src/Category/CategoryPopup.cpp
@@ -48,7 +48,7 @@ int selected_category = 0;
 // ========================================================================================
 void CategoryPopup_DoSelected(HWND hListBox, int idx)
 {
-    int item_data = ListBox_GetItemData(hListBox, idx);
+    int item_data = (int)ListBox_GetItemData(hListBox, idx);
     if (item_data != selected_category) {
         CustomLabel_SetUserDataInt(hCategoryUpdateParentCtl, item_data);
         CustomLabel_SetText(hCategoryUpdateParentCtl, GetCategoryDescription(item_data));
@@ -121,7 +121,7 @@ LRESULT CALLBACK CategoryPopup_ListBox_SubclassProc(
         SendMessage(hWnd, LB_GETITEMRECT, 0, (LPARAM)&rcItem);
         int item_height = (rcItem.bottom - rcItem.top);
         int items_count = ListBox_GetCount(hWnd);
-        int top_index = SendMessage(hWnd, LB_GETTOPINDEX, 0, 0);
+        int top_index = (int)SendMessage(hWnd, LB_GETTOPINDEX, 0, 0);
         int visible_rows = 0;
         int items_per_page = 0;
         int bottom_index = 0;
